Reorganize `scenario`
