+++ app/src/androidTest/java/com/ivy/wallet/compose/scenario/core/DonateTest.kt
@@ -1,4 +1,4 @@
-package com.ivy.wallet.compose.scenario
+package com.ivy.wallet.compose.scenario.core
 
 import com.ivy.wallet.compose.IvyComposeTest
 import dagger.hilt.android.testing.HiltAndroidTest
